requirements.txt
streamlit
pyocr
Pillow